refactored security set biometrics screen
