+++ lib/ui/blocs/security/export.dart
@@ -1,3 +1,2 @@
-export 'biometrics/bloc.dart';
 export 'password/bloc.dart';
 export 'common.dart';
