--- lib/ui/widgets/security_set_biometric/export.dart
@@ -1,3 +1,0 @@
-export 'saving_biometrics_popup_content.dart';
-export 'set_biometric_body.dart';
-export 'set_biometric_title.dart';
